requirements.txt
python-telegram-bot==22.7
fastapi
uvicorn
requests
apscheduler
python-dotenv